+++ browser/create-template/templates/nextjs-site/next.config.mjs
@@ -4,13 +4,7 @@ import path from 'path';
 const nextConfig = {
   // temporary workaround for using symlinked package with turbopack
   outputFileTracingRoot: path.join(import.meta.dirname, '../../../'),
-  webpack: config => {
-    config.resolve.extensionAlias = {
-      '.js': ['.ts', '.tsx', '.js'],
-    };
 
-    return config;
-  },
   experimental: {
     turbo: {},
   },
