+++ browser/create-template/templates/nextjs-site/src/hooks.ts
