+++ browser/create-template/templates/nextjs-site/src/store.ts
